Ticket generator/requirements.txt
Flask==2.3.2
Pillow==9.5.0
pilmoji==2.0.2
requests==2.28.2
fonttools==4.39.3
Pillow==9.5.0
